+++ MANIFEST.in
@@ -1,0 +1,3 @@
+# MANIFEST.in
+
+include src/randdg/*.toml
