+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "randdg"
 version = "1.0.0"
-description = "A lightweight pseudo-random data generator prioritising ease of use and aimed at data scientists, engineer, and analysts."
+description = "A lightweight pseudo-random data generator prioritising ease of use. Aimed at data scientists, engineers, and analysts."
 readme = "README.md"
 authors = [{ name = "Samuel Cook", email = "medians_coned_0j@icloud.com" }]
 license = { file = "LICENSE" }
@@ -17,7 +17,8 @@ classifiers = [
     "Programming Language :: Python :: 3",
 ]
 keywords = ["data", "random", "science", "table"]
-dependencies = ["re", "string", "random", "itertools", "time", "datetime"]
+#dependencies = ["string", "random", "itertools", "time", "datetime"]
+dependencies = []
 requires-python = ">=3.9"
 
 [project.optional-dependencies]
@@ -27,21 +28,21 @@ test = ["pytest"]
 Homepage = "https://github.com/SamuelBCook/randdg"
 
 [project.scripts]
-realpython = "randdg.__main__:main"
+realpython = "src.randdg.__main__:main"
 
-[tool.bumpver]
-current_version = "1.0.0"
-version_pattern = "MAJOR.MINOR.PATCH"
-commit_message = "bump version {old_version} -> {new_version}"
-tag_message = "{new_version}"
-tag_scope = "default"
-pre_commit_hook = ""
-post_commit_hook = ""
-commit = true
-tag = true
-push = true
+# [tool.bumpver]
+# current_version = "1.0.1"
+# version_pattern = "MAJOR.MINOR.PATCH"
+# commit_message = "bump version {old_version} -> {new_version}"
+# tag_message = "{new_version}"
+# tag_scope = "default"
+# pre_commit_hook = ""
+# post_commit_hook = ""
+# commit = true
+# tag = true
+# push = true
 
-[tool.bumpver.file_patterns]
-"pyproject.toml" = ['current_version = "{version}"']
-"README.md" = ["{version}", "{pep440_version}"]
-"src/randdg/__init__.py" = ["{version}"]
+# [tool.bumpver.file_patterns]
+# "pyproject.toml" = ['current_version = "{version}"']
+# "README.md" = ["{version}", "{pep440_version}"]
+# "src/randdg/__init__.py" = ["{version}"]
