+++ src/randdg/__init__.py
@@ -1,0 +1,4 @@
+# __init__.py
+
+# Version of the randdg package
+__version__ = "1.0.0"
